Исправлен тип данных id в схеме и моделях на Integer

+++ src/main/java/ru/bellintegrator/practice/model/Document.java
@@ -17,7 +17,7 @@ import java.util.Date;
 public class Document {
 
     @Id
-    private Long id;
+    private Integer id;
 
     @Version
     private Integer version;
@@ -37,11 +37,11 @@ public class Document {
     @MapsId
     private User user;
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
@@ -84,4 +84,5 @@ public class Document {
     public void setUser(User user) {
         this.user = user;
     }
+
 }
